Make `-fixfn' more robust at handling floats
This change addreses issue #123. The two new optional parameters to
`-fixfn' allow the caller to specify a custom equality test function,
such as an approximate comparison of floats, and a halt test function,
which can trigger a halt to the fixpoint iteration if it fails to
converge.

The default equality test remains `equal'. The default halt test is a
simple counter up to `-fixfn-max-iterations'. The counter is provided by
the new function `-counter'.

The revised tests illustrate the usage of the new parameters.