Edebug support for threading macros
Given an elisp expression of the form:

    (->> (+ 1 1)
         (+ 2)
         (* 3)
         (format "%s"))

Edebug now knows that the first argument is an expression, so users can
step through it. This is an improvement over the current situation,
where users cannot step through any part of a threaded form.

Note that this debugging behaviour matches thread-first in Emacs 25:
http://git.savannah.gnu.org/cgit/emacs.git/tree/lisp/emacs-lisp/subr-x.el?id=emacs-25.0.95#n49