Add annotations to pure functions. (#199)
Adding the declaration (pure t) enables the byte-compiler to evaluate
calls at compile-time. Adding the declaration (side-effect-free t)
allows the byte-compiler to remove calls entirely if the result is not
used.

For functions that take a predicate as an argument, we cannot know
whether that predicate is pure, so we do not annotate those.

We also don't annotate functions that depend on `-compare-fn`, as that
may change at runtime.

Fixes #198.