Fix setf on emacs23 (#205)
`gv-define-simple-setter` is not defined on Emacs 23, so we need to use
the macro `defsetf` on Emacs 23. `defsetf` is not autoloaded, so we must
require 'cl before using it. We don't need to load 'cl on later Emacs
versions.

However, the whole form is macro-expanded on all Emacs versions. Emacs
24+ assumes that `(defsetf ...)` is a function call because it hasn't
loaded 'cl. This generates byte-compile warnings about unused variables
and nonexistent functions, so silence those.