Ensure ' in docstring code snippets is not replaced
As of Emacs 25.1, `foo' is displayed as ‘foo’. This is controlled by
text-quoting-style, but curved quotes are the default.

This causes problems for docstrings that contain ' in code
snippets. Help buffers show e.g. (funcall 'foo) as (funcall ’foo),
which isn't legal elisp.

Instead, escape the quotes with \=. See substitute-command-keys for
the full docs on \= escaping.