Add `-defun'
`-defun' is like `-lambda', but with destructuring. To implement it, abstract
the arglist and form generation code of `-lambda' into separate functions.

Additional benefits of the refactoring are:

- `-lambda' (and by extension, `-defun') can now accept empty argument lists
- `-lambda' (and `-defun') no longer rebinds symbols (fixing the TODO)

Various declaration forms should be handled properly in `-defun'.

Motivation: when this is merged, `lsp-defun' can be made an alias for `-defun'.