`-lambda': support declare-forms; add `-defmacro'
Refactor `dash--destructure-body' to handle declarations and docstrings itself,
meaning that more code can be shared between `-lambda' and `-defun'.

Refactor `dash--destructure-body' to add signatures to the docstring of BODY if
necessary. This means that docstrings for `-defun' and `-lambda', ... now
generate functions with proper signatures.

Add `-defmacro', which is like `-defun' but as a macro.