`dash--destructure-body': optimize: use `-let*'
`-let*' generates less bytecode than `-let', even if otherwise equivalent,
because the latter is implemented in terms of the former, with many temporaries.
This is exacerbated when using dynamic binding, because all of those variables
also need to be bound dynamically.

Add a TODO note because this is an implementation detail; `-let' needs to be
optimized independently.