`-lambda': don't interpret `declare'
For `lambda', `declare' is ignored. Do the same in `-lambda', allowing a
`-lambda' that starts with `declare' to be debugged.

To implement this, extend `dash--decompose-defun-body' to take an optional
DECLARE? parameter that determines whether `declare' is supported in addition to
`interactive'.

Note that macros can still be `interactive'. This is consistent with
`cl-defmacro', which doesn't put the first `interactive' in `cl-block'.