Fix -doto
The macro was broken until now for anything other than list forms,
due to a mistake in quoting.  Use this as an opportunity to rethink
the macro's behavior in this case.  Instead of using funcall, which
is unusual for a macro, thread the value using ->.  Suggested by
Zach Shaftel <zshaftel@gmail.com>.

* dash.el (-doto): Rewrite in terms of ->.  Fix docstring.  Add
Edebug spec.
(--doto): Rename argument for consistency with -doto.  Expand
docstring.  Add Edebug spec.
* dev/examples.el (-doto): Test fixed behavior of -doto.

* README.md:
* dash.texi: Regenerate docs.

Fixes #333.