Prefer n-ary over unary/binary combinators
This is a rewrite and extension of PR #72 to address issue #306.
For discussion, especially wrt performance, see PRs #72 and #308.

* dash.el (-on, -flip, -not, -orfn, -andfn): Return a variadic
function.  Declare as pure and side-effect-free.
(-rotate-args): New combinator suggested by @vapniks in PR #72.
(-const): Declare as pure and side-effect-free.

* NEWS.md (2.19.0): Announce -rotate-args and variadic combinators.

* dev/examples.el (-partition-after-pred): Fix oddp bug waiting to
happen with negative dividends.
(-cons*): Check that &rest args are safe to mutate.
(-on, -flip, -const, -not, -orfn, -andfn): Extend tests.
(-rotate-args): New test.

* README.md:
* dash.texi: Regenerate docs.