Revert recent threading macro indentation changes
This reverts commit 92562909a763d766fd81b817418c2e05d35d50e0
"* dash.el (-->): Indent with 1 distinguished arg."
of 2021-01-10.

This reverts commit 911ef09a30c0a5211f064e7761f9fc26dbf512b6
"* dash.el (->, ->>): Indent with 1 distinguished arg."
of 2021-03-08.

The Clojure-inspired threading macros -> and ->> are quite popular
and were indented until recently with no distinguished arguments as
per Clojure conventions.  In Elisp practice this is not a problem
since the macros have short names.  For the converse argument
against threading macros such as -some-> that have longer names, see
the discussion in #319 and #321.

This commit reverts relatively recent breaking and controversial
indentation changes that tried to make ->, ->>, and --> more
consistent with the rest of Dash, including -some->, -doto,
etc. which are all indented with a distinguished argument despite
being inspired by Clojure.  For the relevant discussion, see #375.

* dash.el (->, ->>, -->): Indent without a distinguished argument.
(-as->, -some->, -some->>, -fixfn): Reindent.
* NEWS.md (2.19.0): Announce indentation change to -->.